+++ app/models/address.rb
@@ -1,3 +1,5 @@
 class Address < ActiveRecord::Base
+  include Concerns::Emailable
+
   belongs_to :addressable, polymorphic: true
 end