+++ changelog.md
@@ -12,6 +12,7 @@
 * Expands FormSubmission to be emailable
 * Use the project_name config var for the backend template title.
 * Show the list of available email template vars.
+* Make the Address model emailable.
 
 
 0.0.12 - 2016-01-22